feat: Allow negative token values in Hash Lookup
Although this isn't officially in the documentation, I've found that the
`value` property of an `InstructionTextToken` of type `IntegerToken`
already always gives the unsigned interpretation of its value,
regardless of how the user is currently displaying it.

Therefore, this should already do what we want in #13, which is to,
when performing a hash lookup or hunt, interpret the 4 bytes backing a
negative `int32_t` value as a `uint32_t`, and to intepret the 8 bytes
backing a negative `int64_t` value as a `uint64_t`.

We can simply remove the warnings. Add some debug logging here as well,
so that we can get some feedback from the logs if there is an issue.

Closes #13.